% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-nodes.R
\name{add_all_grains}
\alias{add_all_grains}
\title{Merge all grains into graph}
\usage{
add_all_grains(graph_base, grains)
}
\arguments{
\item{graph_base}{A panicle graph produced by `read_panicle()`
to which you haven't added spikelets yet}

\item{grains}{A grains dataframe produced by `read_grains()`}
}
\description{
Merge all grains into graph
}
