% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-nodes.R
\name{add_grain}
\alias{add_grain}
\title{Add a Grain to the Closest Branch}
\usage{
add_grain(panicle_graph, new_vert = numeric(2))
}
\arguments{
\item{panicle_graph}{A panicle `tidygraph` object}

\item{new_vert}{Numeric vector that stores the coordinates
of the new grain in format `c(x, y)`}
}
\description{
Add a Grain to the Closest Branch
}
