% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dot-plot.R
\name{get_base}
\alias{get_base}
\title{Get the node at the base of the panicle}
\usage{
get_base(panicle)
}
\arguments{
\item{panicle}{a panicle graph}
}
\description{
This function returns the rank of the node (vertex) at the
base of the panicle. We assume that the base node
is the node  with type: `Generating` with no
children
}
