% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dot-plot.R
\name{get_generating}
\alias{get_generating}
\title{Identify Start and End Generating node}
\usage{
get_generating(panicle)
}
\arguments{
\item{panicle}{a panicle graph.}
}
\description{
This function select the start generating node
as the one with the highest x coordinate and the
end as the node with the lowest
}
\details{
This approach is prone to error is the panicle
is not orientated from right to left in the
P-TRAP picture.

This function is used and improved by `get_base()`,
which select the base node as the generating
node with no parents
}
