% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{good_panicle}
\alias{good_panicle}
\title{A Panicle Graph}
\format{`IGRAPH` with 129 vertexes and 128 edges}
\source{
Produced by Helene Adam at IRD.
}
\usage{
good_panicle
}
\description{
A Graph Storing Vertexes and Edges for A rice Panicle
}
\keyword{datasets}
