% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle-edges.R
\name{invert_edges}
\alias{invert_edges}
\title{Invert Edges on The Main Axis of a Panicle}
\usage{
invert_edges(panicle_graph, check_before = FALSE)
}
\arguments{
\item{panicle_graph}{a panicle graph}

\item{check_before}{`logical`, defaults to `FALSE` should the
function check if
the edges on the main axis of the panicle are pointing inward
before inverting them?}
}
\description{
Invert Edges on The Main Axis of a Panicle
}
