% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{inverted_panicle}
\alias{inverted_panicle}
\title{A Panicle Graph with Inverted Edges}
\format{`IGRAPH` with 134 vertexes and 133 edges}
\source{
Produced by Helene Adam at IRD.
}
\usage{
inverted_panicle
}
\description{
A Graph Storing Vertexes and Edges for A rice Panicle.
The direction of the edges on the main axis (rachis) is
inverted.
}
\details{
This dataset can be used to illustrate how to invert
the directions of the edges on the main axis with
`invert_edges()`.
}
\keyword{datasets}
