% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dot-plot.R
\name{panicle_tibble}
\alias{panicle_tibble}
\title{Turn a Panicle Graph in a Tibble useful for a dotplot}
\usage{
panicle_tibble(panicle, silently = TRUE)
}
\arguments{
\item{panicle}{a panicle graph.}

\item{silently}{logical, if `FALSE` prints intermediate data
while producing the panicle tibble. Defaults to `TRUE`.}
}
\description{
Turn a Panicle Graph in a Tibble useful for a dotplot
}
