% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic-plots.R
\name{plot_panicle}
\alias{plot_panicle}
\title{A simple function to Plot Panilces}
\usage{
plot_panicle(panicle_graph)
}
\arguments{
\item{panicle_graph}{a panicle graph}
}
\description{
plot_panicle is a simple and convinient wrapper to `ggraph()` to plot and
inspect panicle graphs. For versatility, consider calling
`ggraph()` directly.
}
