% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dot-plot.R
\name{pull_branch}
\alias{pull_branch}
\title{Extract a Branch from a Panicle Graph}
\usage{
pull_branch(panicle, vert)
}
\arguments{
\item{panicle}{A graph of class `igraph`}

\item{vert}{A number, the vertex of `panicle`` from which to start}
}
\description{
This function returns a `igraph` with the branch
developing from a selected node.
}
