% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper-parse-and-merge.R
\name{read_full_panicle}
\alias{read_full_panicle}
\title{Read Graph of Panicle Structure and Seeds}
\usage{
read_full_panicle(base_path = NULL, pr_file, gr_file)
}
\arguments{
\item{base_path}{The path to the panicle files without extension.
Use this parameter only if the names of the panicle structure file
and of the panicle structure files are the same, and they are in the
same folder. The `base_path` parameter and the `pr_file` + `gr_file`
parameters are mutually exclusive. Use the first, or the others.}

\item{pr_file}{The path to the `.ricepr` file,
(which stores the structure of a panicle)}

\item{gr_file}{The path to the `.ricegr` file,
(which stores the coordinates of grains)}
}
\description{
This is a convenient wrapper that read:
- A `.ricepr`
file (which stores the structure of a panicle).
- A a `.ricegr`
file (which stores the coordinates of grains).
And merges them together
}
\details{
If the name of the `.ricepr` and the `.ricegr` is the same,
you can provide this name without extension to the parameter
`base_path`. Otherwise, you can provide the path to the two
files with their extensions to the parameters
`pr_file` and `gr_file` .
}
