% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{read_grain}
\alias{read_grain}
\title{Read Grain Spacial Graph}
\usage{
read_grain(path)
}
\arguments{
\item{path}{Path to the panicle XML file}
}
\description{
Read the XML file that stores grain location
(the P-TRAP output that ends by .ricegr)
}
\details{
And returns a tibble that stores seed
 coordinates
}
