% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{read_grains}
\alias{read_grains}
\title{Read Grain Spatial Graph}
\usage{
read_grains(path)
}
\arguments{
\item{path}{Path to the panicle XML file}
}
\description{
Read the XML file that stores grain location
(the P-TRAP output that ends by .ricegr)
}
\details{
And returns a tibble that stores seed
 coordinates
}
