% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{read_panicle}
\alias{read_panicle}
\title{Read Panicle Spatial Graph}
\usage{
read_panicle(path)
}
\arguments{
\item{path}{Path to the panicle XML file}
}
\description{
Read the XML file that stores panicles architecture
data (the P-TRAP output that ends by .ricepr)
}
\details{
And returns a `tidygraph` in which each node has
spatial coordinates
}
