% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dot-plot.R
\name{make_idline}
\alias{make_idline}
\title{Make a Tibble with the Vertex ID of the Longest Path}
\usage{
make_idline(branch, vert_rank)
}
\arguments{
\item{branch}{an `igraph` object storing data for a
panicle branch.}

\item{vert_rank}{a number indicating the `rank` attribute
that identifies the staring vertex.}
}
\description{
This function takes a branch `igraph` as object and
a starting `vertex` as object, finds the longest path
that starts from that vertex.
}
\details{
The longest path is assumend to be the main branch axis

This function returns a character vector that
stores `type` of the vertexes (Secondary, Spikelet, etc.)
along that axis ordered from the starting vertex
(supplied in `vert_rank`) going outward.
}
