% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dot-plot.R
\name{panicle_tibble}
\alias{panicle_tibble}
\title{Turn a Panicle Graph in a Tibble useful for a dotplot}
\usage{
panicle_tibble(panicle, start, to)
}
\arguments{
\item{panicle}{a panicle graph.}
}
\description{
Turn a Panicle Graph in a Tibble useful for a dotplot
}
