% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper-parse-and-merge.R
\name{read_full_panicle}
\alias{read_full_panicle}
\title{Read Graph of Panicle Structure and Seeds}
\usage{
read_full_panicle(pr_file, gr_file)
}
\arguments{
\item{pr_file}{The path to the `.ricepr` file,
(which stores the structure of a panicle)}

\item{gr_file}{The path to the `.ricegr` file,
(which stores the coordinates of grains)}
}
\description{
This is a convenient wrapper that read:
- A `.ricepr`
file (which stores the structure of a panicle).
- A a `.ricegr`
file (which stores the coordinates of grains).
}
\details{
And merges them together
}
